* 'master' of github.com:cpfair/elbbep:
  fixed int vs size_t warning
  removed unused #include "platform.h"
  The magic number 4 has been deposed

+++ runtime/rtl.c
@@ -1,6 +1,5 @@
 #include "rtl.h"
 #include "pebble.h"
-#include "platform.h"
 #include "rtl_ranges.h"
 #include "utf8.h"
 #define SWAP(a, b)                                                             \
@@ -51,7 +50,7 @@ static void apply_swaps(char *ch) {
   // RTL_SWAPS is a set of pairs of chars to be exchanged if they fall in an RTL
   // block.
   char val = *ch;
-  for (int i = 0; i < sizeof(RTL_SWAPS); ++i) {
+  for (size_t i = 0; i < sizeof(RTL_SWAPS); ++i) {
     if (val == RTL_SWAPS[i]) {
       *ch = RTL_SWAPS[i ^ 1];
       return;
